API_WEBAPP/requirements.txt
Werkzeug>=3.0.0
# Static analysis
androguard>=4.1.2
apkutils2>=1.0.0
Pillow>=10.0.0
imagehash>=4.3.1
cryptography>=42.0.0
